Section 11 is over for now so sod it.
I did the rest of the lectures and videos. I gained very little from it, in the end. I feel hollow and unsatisfied. I will now check out some of the section 12 stuff. Then I'll be set up to work on creating Ahoy Matey, as for now it's just Network and LAN crap, which I understand very, very little of.