Ahoy Matey - Ship Controls improved
I have made it so that the ship turns like a 17th century ship does. I also made it so that the ship could not reverse because that made no sense.
It also works over LAN and the two audio listner error has been fixed.
The next step is to make the sky box not horrible to look at when turning the ship or just swap it out for another. Then I'll create, animate and code the cannon balls.