Ahoy Matey - I created a problem
I added a smoke trail to the cannon balls, which caused a problem that I spent most of the day working on until I realised that I just needed to remove the smoke trail. I thentested the cannon ball shooting over LAN  to discover that when the client  presses fire 1 or 2, the host ship shoots instead and when the host shoots, nothing is visibly seen to be happening but cannon balls are created, just not seen.

I think I need work on making them visible and the correct player controlling them but right now I'm tiredand I havea headache so I'll try and do something tomorrow before work and then continue again on Saturday.