Ahoy Matey - I am giving up on this game.
I'm never working on this game again.
I have not got the knowledge with LAN  and Unity to fix the issues and I do not have the will or the time to re do the entire game from scratch in vain hope that I will be able to get it to work properly over LAN.
I am ending this and the course. NOW FUCK OFF ALREADY!